优化MIUI屏蔽更新功能 支持magisk

+++ app/src/main/assets/kr-script/miui/miui_update_set.sh
@@ -1,22 +1,45 @@
-#!/system/bin/sh
+file="/system/etc/hosts"
 
-source ./kr-script/common/mount.sh
-mount_all
+if [[ "$MAGISK_PATH" != "" ]]; then
+    # Magisk 替换
 
-echo '屏蔽MIUI在线更新下载地址(需要解锁System分区)...'
+    if [[ -f $MAGISK_PATH$file ]]; then
+        mkdir -p $MAGISK_PATH/system/etc
+        cp $file $MAGISK_PATH$file
+    fi
 
-path="/system/etc/hosts"
-$BUSYBOX sed '/127.0.0.1\ \ \ \ \ \ \ update.miui.com/'d $path > /cache/hosts
-
-if [[ ! $state = 1 ]]; then
-    $BUSYBOX sed -i '$a127.0.0.1\ \ \ \ \ \ \ update.miui.com' /cache/hosts
+    if [[ $state == 1 ]]; then
+        # 恢复更新 移除规则
+        $BUSYBOX sed -i '/127.0.0.1\ \ \ \ \ \ \ update.miui.com/'d $MAGISK_PATH$file
+    else
+        # 屏蔽更新 添加规则
+        $BUSYBOX sed -i '$a127.0.0.1\ \ \ \ \ \ \ update.miui.com' $MAGISK_PATH$file
+    fi
     pm clear com.android.updater 2> /dev/null
-    echo '已添加“127.0.0.1        update.miui.com”到hosts'
-fi;
 
-cp /cache/hosts $path
-chmod 0755 $path
-rm /cache/hosts
-sync
+    echo '此操作需要重启手机方可生效！'
+else
+    # 非 Magisk 替换
 
-echo '可能需要重启手机才会生效！'
+    source ./kr-script/common/mount.sh
+    mount_all
+
+    echo '屏蔽MIUI在线更新下载地址(需要解锁System分区)...'
+
+    path="/system/etc/hosts"
+    $BUSYBOX sed '/127.0.0.1\ \ \ \ \ \ \ update.miui.com/'d $path > /cache/hosts
+
+    if [[ ! $state = 1 ]]; then
+        $BUSYBOX sed -i '$a127.0.0.1\ \ \ \ \ \ \ update.miui.com' /cache/hosts
+        pm clear com.android.updater 2> /dev/null
+        echo '已添加“127.0.0.1        update.miui.com”到hosts'
+    fi;
+
+    cp /cache/hosts $path
+    chmod 0755 $path
+    rm /cache/hosts
+    sync
+
+    echo '可能需要重启手机才会生效！'
+fi
+
